# feat(file sink): supports input based on encoding type