--- a/changelog.d/21704_file_sink_supports_encoding_input.feature.md
+++ b/changelog.d/21704_file_sink_supports_encoding_input.feature.md
@@ -1,0 +1,3 @@
+The file sink now supports any input event type that the configured encoding supports. It previously only supported log events.
+
+authors: nionata