@pront 2024-11-15
> > Hi @nionata, just checking if you wanted a review on this or it's purely a draft for your own purposes.
> 
> Hello @pront! I would like to get this reviewed and merged. I have it as draft because I wanted to add a test. Feel free to let me know early if this is something we could get merged or not.

Makes sense, no rush 👍 
We will definitely benefit from a test. Ping me when it's ready for review.