@nionata 2024-11-21
> Thankss @nionata, this looks good! I have only one concern, with this PR we also accept traces. Is it possible to add a basic test for traces as well? Just to make sure we don't falsely advertise trace support. If we find a problem, we can just exclude traces.

Agreed. Pushed not supported and supported trace tests.