@pront 2024-11-22
Ah, please add a changelog fragment. See readme and examples here: https://github.com/vectordotdev/vector/tree/master/changelog.d