@nionata 2024-11-22
```bash
/git/vectordotdev/vector# cargo clippy --all-targets
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.93s
```